avoid comparing versions when sorting
a bug in the way we cache version objects can result in
a cached object returned from versions.VersionFromString()
to be changed by another recipient of that version object.
Often the change is adding timestamps.  By chance, the versions
in the job sets got timestamps.  When version caching was added
to the local database, these timestamps are no longer added to
cached version objects.