Added memusage(), and fixed memsize() to use memusage().
/proc/self is a symlink to /proc/%s % (os.getpid(), ) and it's easier to use.
Also, reading the 11th line from /proc/self/status is not a portable way to
retrieve VmSize; /proc/self/statm is easier to parse and has the size of the
shared memory too.
Now, there is a ton of information on the net about what each field means and
how to analyze memory usage, and they all seem to have a slightly different
spin to it. memusage() returns all the fields from /proc/self/statm, if
not a correct absolute reading it should help with relative readings (i.e. how
much did VmRss increase from one point to another).