Added support for authenticated proxies. (CNY-1687)
Fixed the socket.error we synthesize when talking to a Conary proxy to use a
negative value, denoting it's one of our errors.Added support for
authenticated proxies. (CNY-1687)

Fixed the socket.error we synthesize when talking to a Conary proxy to use a
negative value, denoting it's one of our errors.Added support for
authenticated proxies. (CNY-1687)

Fixed the socket.error we synthesize when talking to a Conary proxy to use a
negative value, denoting it's one of our errors.