Allow other Python threads to run when calling sqlite3_prepare() (CNY-2568)
    * A bug that occasionally caused a thread deadlock when multiple
      threads access the local system Conary database concurrently has
      been fixed. (CNY-2586)

This was caused by two threads accessing the database at the same
time: the update thread and the changeset download thread.  The update
thread is in the process of adding to the TroveInfo table.  The cs
download thread is getting some information about existing troves in
the database, and is accessing the database for the first time.

sqlite3_prepare called by the cs download thread needs to load the
database schema.  This requires a SHARED_LOCK on the database.  It is
holding the Python Interpreter Lock (PIL) while it is trying to get
the database lock.

The update Python thread is in the process of updating the database,
so the cs download thread can't obtain the lock.  In fact, the update
Python thread has just called sqlite3_step and is now trying to
re-acquire the PIL.  It can't relinquish its database lock until the
SQL statement is finished executing.  Since it can't acquire the PIL
(because the cs download thread holds it), we deadlock.

The fix is to allow other Python threads to execute while waiting for
sqlite3_prepare() to return.  This lets the other Python thread run
and relinquish its database lock.