packagepolicy: add bootstrapRubyLibs keyword to Provides and Requires
When bootstrapping ruby needs to know where its libraries are located,
which is dependent on the api version (which can differ from the ruby
program version). This change adds a keyword argument to the Provides
and Requires policies, bootstrapRubyLibs, that lets the packager pass in
a list of paths to ruby libraries.