Simplify RegularFile.restore() considerably
This pulls some syscall code out of sha1Uncompress that was also
duplicated in restore(). The overhead of setting up a try: block in
Python is trivial next to the syscall itself so there's no reason to
push the file manipulation code into C when the bulk of the time is
spent doing uncompress/digest operations.

Also removes an unnecessary file copy when updating hardlinks. It seemed
to have been done that way to fix a bug without touching sha1Uncompress,
but making a hardlink "copy" should work perfectly fine.

Finally, the pure-Python implementation of sha1Uncompress is removed
because it serves no purpose when RegularFile.restore() already has a
more general fallback implementation.