Relax testPtrFileChangesContent to fit a more realistic use case
Previously the test mangled a file in place during an update. Now it
replaces it with a new file using rename(). The original scenario that
the test was developed to validate seemed to be:

1. A .pyc and .pyo file have same content
2. In the update changeset .pyo points to .pyc
3. Another python process imports the new .py in the middle of the
   update which causes a new .pyc to be written out
4. Conary attempts to install the .pyo by copying the .pyc but it's been
   replaced with a different one and the SHA-1 check fails

The changes to sha1Uncompress in the commit preceding this one mean that
instead of duplicating the .pyc immediately, we instead make a hardlink
and later copy the pointer files from that.  This means that mangling
the first file will still cause a SHA-1 error, but replacing it with a
new file (as Python does) will work because it breaks the hardlink. This
seems like an acceptable tradeoff.