Fix conary rdiff failing for misisng sub-troves and capsules (APPENG-3093)
If a sub-trove is removed in the newer version of a changeset, we need
to handle generating the diff differently. Essentially, we rely solely
on the infomration in the old trove. Also, if a capsule has changed,
then we can only say that the files differ, we cannot generate a diff.