netclient: fix getChangeSetSize error on trove removal (APPENG-3272)
We were naively retrieving a repository based on the new version of the
first job in the job list. If the first job removed a trove, this meant
conary tried to talk to a repo named None. Instead, check if the new
version is None, and if it is use the old version.