Resolve many test suite failures on Fedora 22
Many tests have been updated to correctly handle "usrmove" and
to function correctly whether or not they are running on a
system with that change.

Several tests are updated not to depend on the uucp and daemon
users existing, and more particularly existing with particular
IDs.

A few tests checked explicitly for Python 2.6 to determine
expected behavior relative to older versions of Python; those
tests now also check for 2.7.

Perl tests were changed to adapt to changes in the perl-CGI
package; CGI::Apache no longer exists in perl-CGI, so change
instead to CGI::Util which does.

The python tarfile module changed the error returned for an
empty archive.  The "Name or service not known" error message
was changed to "No address associated with hostname".

Use the documented context manager for testing warnings.warn()
to work correctly in Python 2.7.

Handle variant behavior in the internal and undocumented
urllib.splituser() introducing a bug fix between Python
2.6 and 2.7.

Several tests were modified to have more useful error messages
or be otherwise easier to debug future problems.

Change-Id: I203506b5ce4c082893ad8845f98ad99a41c09cd4